refactor: core디렉터리 삭제 및 common디렉터리 하위로 SelectBox 컴포넌트 이동
